fix: 巻き戻しの自動レビュー指摘を消化する
**無効の付与が復活する穴を塞ぐ。** 受注 A の付与を受注 B が使い切ると、A の巻き戻しは
打ち消す対象を持たず記録だけが残る。その後 B を巻き戻すと引き当ての逆転が A のロットへ
量を返し、A は二度目の巻き戻しを受け付けないため、無効にしたはずの付与が使える残高として
復活していた。打ち消しの対象を「本受注の付与」から「本受注の付与、および逆転で量が戻る
先のうちその受注が既に巻き戻されているもの」へ広げる。

**巻き戻し済みの受注へは伝票を再発行しない。** 発行しても申領は拒まれるので、店舗は成功
したと告げられたうえで使えない QR を客へ見せることになる。判定は controller に置く —
再発行のサービスが台帳へ依存を持たないこと自体が「無効化は台帳へ波及しない」（ADR 0009）の
構造的な証跡であり、現況の読み口が 2 サービスを controller で組み立てているのと同じ扱い。
述語の側は先に現店舗の受注として引く。引かないと、他店舗の受注 ID に対して巻き戻し済みかが
応答の違い（400 か 404 か）から読み取れる。

**単一の宛先という言い切りをやめる。** 同じ受注が無効化と再申領で別の会員へ帰属し直して
いれば付与は複数の会員に跨りうるが、下見は現に帰属している 1 人の会員コードしか返さない。
コードの意味を「現帰属」と明示し、画面は過去の帰属分もまとめて取り消されることを名乗る。

**合計を long にする。** 1 受注に複数の付与が積まれた合計は int を超えうる（残高や受注ごとの
付与合計と同じ理由）。1 件の未消費分は自身の加算量が上限なので int のまま。

**再試行が受注の取り直しも起こすようにする。** 下見だけを取り直すと、落ちたのが受注側だった
ときに押しても何も直らない。

+++ backend/src/main/java/com/kizuna/order/api/dto/OrderPointRollbackPreviewResponse.java
@@ -6,12 +6,13 @@ package com.kizuna.order.api.dto;
  * <p>会員の残高も載せない。この受注の授受に限って示すのであって、店舗へ会員の資産を渡す口ではない（ADR 0006）。
  *
  * @param alreadyRolledBack 既に巻き戻し済みか。真なら二度目の要求は 409 になる
- * @param memberCode この受注が現に帰属している会員のコード。帰属していなければ null
+ * @param memberCode この受注が<b>現に帰属している</b>会員のコード。帰属していなければ null。打ち消す付与はこの会員のものとは限らない —
+ *     同じ受注が無効化と再申領で別の会員へ帰属し直していれば、前の会員の付与も対象になる
  * @param cancellablePoints 取消で無効化される付与の未消費残の合計
  * @param reversibleUsedPoints 逆転で元のロットへ返る利用の合計
  */
 public record OrderPointRollbackPreviewResponse(
     boolean alreadyRolledBack,
     String memberCode,
-    int cancellablePoints,
-    int reversibleUsedPoints) {}
+    long cancellablePoints,
+    long reversibleUsedPoints) {}
